requirements.txt
# Core dependencies - Updated to latest secure versions
# NOTE: Flask is retained for legacy test compatibility. FastAPI is the primary backend.
Flask==3.0.0
numpy>=1.24.0

# FastAPI Backend (Primary)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
email-validator>=2.0.0
python-multipart>=0.0.6
httpx>=0.26.0
pandas>=2.0.0
requests==2.31.0
Werkzeug==3.0.0
Jinja2==3.1.2
itsdangerous>=2.1.0
click>=8.1.0

# WebSocket support
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
gevent==23.9.1
gevent-websocket==0.10.1

# Security enhancements
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
cryptography>=41.0.0
python-dotenv==1.0.0
PyJWT==2.8.0

# GraphQL support (Compatible versions)
graphene>=3.0,<3.3
graphene-sqlalchemy==3.0.0b4
graphql-core>=3.0,<4.0

# Database (Compatible with graphene-sqlalchemy)
SQLAlchemy>=1.4.0,<2.0.0
alembic>=1.12.0

# Data acquisition
geopy==2.3.0

# Visualization
plotly==5.17.0
matplotlib>=3.7.0

# Scientific computing
scipy>=1.11.0
scikit-learn>=1.3.0

# Additional scientific libraries for quantum platform
seaborn>=0.12.0
networkx>=3.0
joblib>=1.3.0

# Image processing (optional)
Pillow>=10.0.0

# Text processing (optional)  
nltk>=3.8
spacy>=3.6.0

# Audio processing (optional)
librosa>=0.10.0

# Quantum computing (optional dependencies)
qiskit>=0.45.0
pennylane>=0.32.0
autoray>=0.6.0,<0.7.0
qiskit-aer>=0.12.0
qiskit-machine-learning>=0.7.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-html>=4.1.0
pytest-xdist>=3.5.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# End-to-end testing with Playwright
playwright>=1.40.0
pytest-playwright>=0.4.3

# Error monitoring and performance tracking
sentry-sdk[flask]>=1.38.0
psutil>=5.9.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# WebSocket support for FastAPI
websockets>=12.0

# Additional testing utilities
coverage>=7.3.0
faker>=20.1.0
hypothesis>=6.88.0
locust>=2.17.0

# Production deployment
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
redis>=4.6.0
celery>=5.3.0

# Monitoring and logging
structlog>=23.1.0
prometheus-client>=0.18.0
prometheus-flask-exporter>=0.23.0